pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghsettings"
version = "0.0.1"
description = "Configure GitHub settings through the CLI"
requires-python = ">=3.7"
dependencies = [
    "PyGithub>=2.5.0",
    "halo==0.0.31",
    "urllib3==1.26.18",
]
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
license = "MIT"

[project.scripts]
ghsettings = "ghsettings.ghsettings:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ghsettings"]
